+++ prisma/migrations/20241011194527_add_leaderboard_per_participant_goals/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "BoardParticipant" ADD COLUMN     "goal" JSONB;