+++ prisma/migrations/20241011195328_add_explicit_individual_goal_mode_flag/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Board" ADD COLUMN     "individualGoalMode" BOOLEAN NOT NULL DEFAULT false;