+++ src/components/board/BoardIndividualProgressChart.vue
@@ -4,26 +4,28 @@ import { computed, defineProps } from 'vue';
 import type { Board, ParticipantWithTallies } from 'src/lib/api/board.ts';
 
 import { normalizeTallies, accumulateTallies, listEachDayOfData } from '../chart/chart-functions.ts';
-import { TallyMeasure } from 'server/lib/models/tally.ts';
-import { formatCount } from 'src/lib/tally.ts';
 
 import PlotProgressChart from 'src/components/chart/PlotProgressChart.vue';
 import { densifyTallies } from '../chart/chart-functions.ts';
+import { formatPercent } from 'src/lib/number.ts';
 
 const props = defineProps<{
   board: Board;
   participants: ParticipantWithTallies[];
-  measure: TallyMeasure;
 }>();
 
-const participants = computed(() => {
-  return props.participants.filter(participant => participant.tallies.filter(tally => tally.measure === props.measure).length > 0);
+const isFundraiserMode = computed(() => {
+  return props.board.fundraiserMode && !props.board.individualGoalMode;
 });
+
+const participants = computed(() => {
+  return props.participants.filter(participant => participant.tallies.filter(tally => tally.measure === participant.goal.measure).length > 0);
+})
 
 const chartData = computed(() => {
   const [ earliestDate, latestDate ] = participants.value
     .flatMap(participant => participant.tallies
-      .filter(tally => tally.measure === props.measure)
+      .filter(tally => tally.measure === participant.goal.measure)
       .map(tally => tally.date)
     )
     .reduce(([earliest, latest], date) => {
@@ -40,30 +42,30 @@ const chartData = computed(() => {
     par: null,
   };
 
-  // add a par if needed
-  if(props.measure in props.board.goal) {
-    const goalCount = props.board.goal[props.measure];
+  // add a par line from 0 to 100%
+  const goalCount = 100;
 
-    const parData = eachDay.map((date, ix) => ({
-      series: 'Par',
-      date,
-      value: props.board.endDate === null ?
-        goalCount :
-        (ix === eachDay.length - 1) ? goalCount : Math.ceil((goalCount / eachDay.length) * (ix + 1)),
-    }));
-    data.par = parData;
-  }
+  const parData = eachDay.map((date, ix) => ({
+    series: 'Par',
+    date,
+    value: props.board.endDate === null ?
+      goalCount :
+      (ix === eachDay.length - 1) ? goalCount : Math.ceil((goalCount / eachDay.length) * (ix + 1)),
+  }));
+  data.par = parData;
 
   data.tallies = participants.value.flatMap(participant => {
-    const tallies = participant.tallies.filter(tally => tally.measure === props.measure);
+    const tallies = participant.tallies.filter(tally => tally.measure === participant.goal.measure);
     const normalizedTallies = normalizeTallies(tallies);
-    const densifiedTallies = props.board.fundraiserMode ? densifyTallies(normalizedTallies, eachDayOfData) : normalizedTallies;
+    const densifiedTallies = isFundraiserMode.value ? densifyTallies(normalizedTallies, eachDayOfData) : normalizedTallies;
     const accumulatedTallies = accumulateTallies(densifiedTallies);
 
     const participantTallyData = accumulatedTallies.map(tally => ({
       series: participant.displayName,
       date: tally.date,
-      value: tally.value,
+      value: +formatPercent(tally.value, participant.goal.count),
+      raw: tally.value,
+      measure: participant.goal.measure,
     }));
     return participantTallyData;
   });
@@ -77,13 +79,13 @@ const chartData = computed(() => {
   <PlotProgressChart
     :data="chartData.tallies"
     :par="chartData.par"
-    :is-stacked="props.board.fundraiserMode"
+    :is-stacked="isFundraiserMode"
     :config="{
-      measureHint: props.measure,
+      measureHint: 'percent',
       seriesTitle: 'Participant',
       showLegend: true,
     }"
-    :value-format-fn="d => formatCount(d, props.measure)"
+    :value-format-fn="d => `${d}%`"
   />
 </template>