+++ src/components/board/EditBoardParticipationForm.vue
@@ -12,12 +12,18 @@ tagStore.populate();
 
 import { z } from 'zod';
 import { useValidation } from 'src/lib/form.ts';
+import { TALLY_MEASURE } from 'server/lib/models/tally.ts';
+import type { NonEmptyArray } from 'server/lib/validators.ts';
+import { TALLY_MEASURE_INFO } from 'src/lib/tally.ts';
+import type { ParticipantGoal } from 'server/lib/models/board';
 
 import { joinBoard, BoardParticipantPayload, Board, ExtendedBoardParticipant } from 'src/lib/api/board.ts';
 
 import MultiSelect from 'primevue/multiselect';
+import Dropdown from 'primevue/dropdown';
 import TbForm from 'src/components/form/TbForm.vue';
 import FieldWrapper from 'src/components/form/FieldWrapper.vue';
+import TallyCountInput from '../tally/TallyCountInput.vue';
 // import TbTag from 'src/components/tag/TbTag.vue';
 
 const props = withDefaults(defineProps<{
@@ -33,16 +39,33 @@ const isJoin = computed(() => {
 });
 
 const formModel = reactive({
+  measure: isJoin.value ? TALLY_MEASURE.WORD : (props.participant.goal as ParticipantGoal)?.measure ?? TALLY_MEASURE.WORD,
+  count: isJoin.value ? null : (props.participant.goal as ParticipantGoal)?.count ?? null,
   works: isJoin.value ? [] : props.participant.worksIncluded.map(work => work.id),
   tags: isJoin.value ? [] : props.participant.tagsIncluded.map(tag => tag.id),
 });
 
 const validations = z.object({
+  measure: z.enum(Object.values(TALLY_MEASURE) as NonEmptyArray<string>).nullable(),
+  count: z
+      .number({ invalid_type_error: 'Please enter a value.' }).int({ message: 'Please enter a whole number.' }).nullable()
+      .refine(v => props.board.individualGoalMode ? v !== null : true, { message: 'Please input your goal for this leaderboard.'}),
   works: z.array(z.number({ invalid_type_error: 'Please select only valid projects.' }).int({ message: 'Please select only valid projects.' }).positive({ message: 'Please select only valid projects.' })),
   tags: z.array(z.number({ invalid_type_error: 'Please select only valid tags.' }).int({ message: 'Please select only valid tags.' }).positive({ message: 'Please select only valid tags.' })),
 });
 
 const { ruleFor, validate, isValid, formData } = useValidation(validations, formModel);
+
+const measureOptions = computed(() => {
+  return Object.values(TALLY_MEASURE).map(measure => ({
+    id: measure,
+    label: TALLY_MEASURE_INFO[measure].label.plural,
+  }));
+});
+
+const onMeasureChange = function() {
+  formModel.count = null;
+}
 
 const isLoading = ref<boolean>(false);
 const successMessage = ref<string | null>(null);
@@ -55,7 +78,16 @@ async function handleSubmit() {
 
   try {
     const data = formData();
-    const updated = await joinBoard(props.board.uuid, data as BoardParticipantPayload);
+    const payload: BoardParticipantPayload = {
+      goal: props.board.individualGoalMode ? {
+        measure: data.measure,
+        count: data.count,
+      } : null,
+      works: data.works,
+      tags: data.tags,
+    };
+
+    const updated = await joinBoard(props.board.uuid, payload);
 
     emit('participation:edit', { participation: updated });
     successMessage.value = `You have ${isJoin.value ? 'joined' : 'changed your selections for'} ${props.board.title}!`;
@@ -84,6 +116,41 @@ async function handleSubmit() {
     @submit="validate() && handleSubmit()"
     @cancel="emit('formCancel')"
   >
+    <FieldWrapper
+      v-if="props.board.individualGoalMode"
+      for="board-form-goal-count"
+      label="Your individual goal"
+      :rule="ruleFor('count')"
+      required
+      help="This leaderboard requires every participant to set their own individual goal."
+    >
+      <template #default="{ onUpdate, isFieldValid }">
+        <div class="count-fieldset-container flex gap-2">
+          <div class="count-fieldset-measure flex-none">
+            <Dropdown
+              id="board-form-goal-measure"
+              v-model="formModel.measure"
+              :options="measureOptions"
+              option-label="label"
+              option-value="id"
+              @change="onMeasureChange"
+            />
+          </div>
+          <div class="count-fieldset-count flex-auto">
+            <TallyCountInput
+              id="board-form-goal-count"
+              v-model="formModel.count"
+              :measure="formModel.measure"
+              :invalid="!isFieldValid"
+              @update:model-value="onUpdate"
+            />
+          </div>
+        </div>
+      </template>
+    </FieldWrapper>
+    <div class="">
+      Select which progress updates you want to include on this leaderboard. You can filter by project, tag, or both.
+    </div>
     <FieldWrapper
       for="board-form-works"
       label="Projects to include"